feat(search): extract useShowResults and add tests
