+++ package.json
@@ -188,7 +188,6 @@
     "@testing-library/react": "^12.0.0",
     "@testing-library/react-hooks": "^3.4.2",
     "@testing-library/react-native": "7.1.0",
-    "@testing-library/user-event": "^14.2.0",
     "@types/amplitude-js": "^8.9.6",
     "@types/contentful-resolve-response": "^0.1.30",
     "@types/geojson": "^7946.0.7",
