+++ src/features/search/pages/Search.tsx
@@ -1,5 +1,5 @@
 import { useNavigation, useRoute } from '@react-navigation/native'
-import React, { useEffect, useRef, useState } from 'react'
+import React, { useEffect } from 'react'
 import styled from 'styled-components/native'
 import { v4 as uuidv4 } from 'uuid'
 
@@ -7,40 +7,9 @@ import { useAppSettings } from 'features/auth/settings'
 import { UseNavigationType, UseRouteType } from 'features/navigation/RootNavigator'
 import { SearchHeader, SearchLandingPage, SearchResults } from 'features/search/components'
 import { useSearch } from 'features/search/pages/SearchWrapper'
+import { useShowResults } from 'features/search/pages/useShowResults'
 import { useKeyboardAdjust } from 'ui/components/keyboard/useKeyboardAdjust'
 import { Form } from 'ui/web/form/Form'
-
-import { useSearchResults } from './useSearchResults'
-
-export const useShowResults = () => {
-  const timer = useRef<number>()
-  const { searchState } = useSearch()
-  const [showResults, setShowResults] = useState<boolean>(searchState.showResults)
-  const { isLoading } = useSearchResults()
-
-  useEffect(() => {
-    if (searchState.showResults) {
-      if (!isLoading) {
-        setShowResults(true)
-      } else {
-        // For most networks, 20ms is enough time to fetch the results fom Algolia
-        // In this case we can avoid displaying the placeholders
-        timer.current = globalThis.setTimeout(() => {
-          setShowResults(true)
-        }, 20)
-      }
-    } else {
-      setShowResults(false)
-    }
-    return () => {
-      if (timer.current) {
-        clearTimeout(timer.current)
-      }
-    }
-  }, [searchState.showResults, isLoading])
-
-  return showResults
-}
 
 export function Search() {
   useKeyboardAdjust()
